Update devicemapper to use u128-based Bytes

+++ src/engine/strat_engine/metadata/bda.rs
@@ -155,7 +155,7 @@ mod tests {
         /// Initialize a BDA.
         /// Verify that the last update time is None.
         fn empty_bda(ref sh in static_header_strategy()) {
-            let buf_size = convert_test!(*sh.mda_size.bda_size().sectors().bytes(), u64, usize);
+            let buf_size = convert_test!(*sh.mda_size.bda_size().sectors().bytes(), u128, usize);
             let mut buf = Cursor::new(vec![0; buf_size]);
             let bda = BDA::initialize(
                 &mut buf,
@@ -181,7 +181,7 @@ mod tests {
             0;
             convert_test!(
                 *sh.blkdev_size.sectors().bytes(),
-                u64,
+                u128,
                 usize
             )
         ]);
@@ -202,7 +202,7 @@ mod tests {
             0;
             convert_test!(
                 *sh.blkdev_size.sectors().bytes(),
-                u64,
+                u128,
                 usize
             )
         ]);
@@ -234,7 +234,7 @@ mod tests {
             ref state in vec(num::u8::ANY, 1..100),
             ref next_state in vec(num::u8::ANY, 1..100)
         ) {
-            let buf_size = convert_test!(*sh.mda_size.bda_size().sectors().bytes(), u64, usize);
+            let buf_size = convert_test!(*sh.mda_size.bda_size().sectors().bytes(), u128, usize);
             let mut buf = Cursor::new(vec![0; buf_size]);
             let mut bda = BDA::initialize(
                 &mut buf,
